all done

+++ includes/libftprintf.h
@@ -6,7 +6,7 @@
 /*   By: jthomas <jthomas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/16 03:45:00 by jthomas           #+#    #+#             */
-/*   Updated: 2023/03/01 18:33:35 by jthomas          ###   ########.fr       */
+/*   Updated: 2023/03/05 16:36:56 by jthomas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ typedef struct s_list
 int		ft_printf(const char *format, ...);
 int		ft_printflag(int fd, const char *format, va_list args, int *error);
 int		ft_puthex_fd(unsigned int n, char format, int fd, int *error);
-int		ft_putunsigned_fd(unsigned int n, int fd, int *error);
+int		ft_putunsigned_fd(long long int n, int fd, int *error);
 int		ft_putunsignedhex_fd(unsigned long n, int fd, int *error);
 int		ft_putchar_fd(char c, int fd, int *error);
 int		ft_putstr_fd(char *s, int fd, int *error);
