+++ src/ft_printflag.c
@@ -6,7 +6,7 @@
 /*   By: jthomas <jthomas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/06 23:46:52 by jthomas           #+#    #+#             */
-/*   Updated: 2023/03/01 18:56:51 by jthomas          ###   ########.fr       */
+/*   Updated: 2023/03/05 16:49:46 by jthomas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@ int	ft_print(int fd, const char format, va_list args, int *error)
 	if (format == 'd' || format == 'i')
 		return (ft_putnbr_fd(va_arg(args, int), fd, error));
 	if (format == 'u')
-		return (ft_putunsigned_fd(va_arg(args, int), fd, error));
+		return (ft_putunsigned_fd(va_arg(args, long long int), fd, error));
 	if (format == 'x' || format == 'X')
 		return (ft_puthex_fd(va_arg(args, int), format, fd, error));
 	if (format == '%' || format == '#')
@@ -32,7 +32,8 @@ int	ft_print(int fd, const char format, va_list args, int *error)
 	if (format == 'p')
 	{
 		res += ft_putstr_fd("0x", fd, error);
-		res += ft_putunsignedhex_fd(va_arg(args, unsigned long long), fd, error);
+		res += ft_putunsignedhex_fd(
+				va_arg(args, unsigned long long), fd, error);
 	}
 	return (res);
 }
